Standardise the use of baseURL util across all components for consistency
The baseURL simply stores the value of the API URL environment variable,
and is used when making requests to the API, rather than having to type
the full domain in code. This makes for better consistenct and easier
maintenance. Changing the API domain will not require a change to the
codebase, but only the REACT_APP_API_URL environment variable.