Fix Handover conditional logic in Shift Details panel
The conditional logic allowing the submission of handover should not
focus on whether the patient has a next shift, but whether the currently
selected shift is the last or next-to-last.

To add handover, the selected shift should be in progess, OR the
last shift in the list of shifts, OR the second-to-last shift in
the list of shifts with the last shift still uncommenced.