Refactor conditonal rendering of Shift Details and child components
Rather than having nested ternary operator in the component return
statement, components in the Shift Details panel now use a renderContent
function that handles conditional logic and determines the content
to render for a given state.

A utility object (shiftUtils) has been created to check for a range
of states for the selected shift, which can be easily accessed from
the object properties. shiftUtils stores its state in ShiftDetails.jsx
and is passed to child functions as a prop.

It currently stores the following as Boolean values:
  - userIsCarer: Whether the logged-in user is the carer for the shift
  - isLastShift: Whether the shift is the last shift in the list of shifts
  - isPenultimateShift: Whether the shift is next-from-last
  - isPending: Whether the shift is in the future (uncommenced)
  - isInProgress: Whether current time is between shift start and end time
  - hasEnded: Whether current time is after shift end time
  - nexShiftHasStarted: Whether the final shift of client has started
  - isInEditWindow: Whether the shift has ended but is still editable

In future, nextShiftHasStarted would more usefully check the next shift
(adjacent to the selected shift) rather than the last shift of the client.