Fix error when opening Shift Details with only one shift
If the client has only one shift in their calendar, nextShift, prevShift,
and penultimateShift will be assigned null in shiftUtils, potentially raising
errors when they are used as the first operand in an && comparison operation
(returning null) that assigns a shiftUtils 'is' or 'has' property (these should
only have Boolean values assigned).

When using nextShift, prevShift, or penultimateShift as the first operand
in an && comparison, they should be cast to a Boolean first to avoid errors.