Improve add/remove carer workflows and data fetching
Coordinators now have a button in the Invite Carer dialog to
quickly add themselves as the carer for a client without going
through the email verification flow.

The Care Team section of the Calendar component and the Care Team List
modal now update immediately when a carer is added to or removed
from a client's care team.

Made adjustments to data fetching functions and state updates
for components that display carers for better consistency between
the database and local (browser) state.