Pre-populate start time when adding a shift via the Upcoming Shift card
Because no selectedDate is set unless a calendar day grid is clicked,
creating a shift from the Upcoming Shift card previously opened a
dialog with an 'Invalid' (empty) shift start time.

Now, the 'start' property of the selectedDate object will be populated with
a time eight hours from the current time if it is empty.

The getYearMonthDay function in shiftUtils now adds 1 to its month variable
before converting to a string, to prevent returning the previous instead of
current month, since the month property of a Date object is zero-indexed.

Also refactored the plusHours and minusHours functions in shiftUtils
to use the setHours and getHours methods to mutate input dates for
cleaner code.