Fix shift edit window logic to prevent premature shift locking
Previously, a shift would be in its 'edit window' on condition that it
was within 8 hours of its ending time and the next shift had not yet started.
The shiftUtils.isInEditWindow property now accounts for the case where the
shift is within its edit window and there is no next shift.