Fix premature shift locking and simplify edit window logic
Shift notes, incident reports, and handover will now only check whether
the shift is in progress or in the edit window to determine whether
they can be edited. The shiftUtils.isInEditWindow logic has been adjusted
to take into account whether the next shift has started, so checking
whether the shift is the penultimate or last shift during the conditional
rednering of note, report, and handover fields is now redundant.