Add PDF upload error handling to shiftController
Uploading the shift notes PDF to cloud file storage has been uncoupled
from saving the shift notes to the database, since a failure to upload
to cloud storage should not result in the loss of entered notes or errors
on the front end.

If PDF upload fails, the shiftNotesPDF field of the shift object is
populated with an empty string.

The download button has also been removed from the Shift Notes tab of the
Shift Details panel if no shiftNotesPDF URL string is present. Instead,
give the user an option to re-attempt PDF conversion and upload of notes
if it failed during the initial process of saving the Shift Notes.