Add Shift Note editing and deletion options
Also added a load state handler to the ShiftNotes panel that passes loading
state down (and back up) through props, allowing the Shift Notes panel,
Shift Notes Form, and Form components to respond to the loading state
of their respective children.

Also added a backend pdfUtils function to delete assets in Cloudinary
and a frontend apiUtils function to update a shift by passing in the
shift id and a JSON body containing the fields to be updated.

The buttons to edit and clear the shift notes will change to 'save' and
'cancel edit mode' buttons respectively when edit mode is entered.