Avoid excessive shiftsOverlap calls if the selected date is in the past
If the selected date is in the past, the getShiftTimeDefaults function
will set the default start to the current time, to avoid searching
through all of the dates between the selected date and the present
for a valid shift timeslot.

Also set eventOrderStrict to true in the CalendarDayGrid FullCalendar
component to prevent shifts that span multiple days from appearing after
later shifts on the subsequent day.

Display modal alerts after the text content for a cleaner layout.