Use the current date when selecting Add Shift from the Calendar sidebar
This resolves a situation where the date used by the Add Shift button
in the sidebar is the currently selected date on the calendar, even
if that date is in the past or future. Now, selecting Add Shift from
the Calendar view sidebar will always default to the current date,
while doing so from the Select Shift By Date dialog will use the
selected date as set by the calendar grid.

Also lifted the ref and getApi method for FullCalendar up from CalendarDayGrid
to Calendar so that they can be passed to other components in the Calendar
view which may need to manipulate the calendar using its API.