Add UI and server route for creating demo accounts
Demo accounts use auto-generated credentials and automatically
log in the user when created, preventing the need for the user
to supply credentials before starting a demo of the app.

Also added an environment variable to the React server indicating
whether the app should be in Demo mode (bool). Front end components
can check the value of process.env.REACT_APP_DEMO and conditionally
alter their output for demo mode.