Fix bugs when cancelling shifts and visiting empty calendar for a new client
The getShiftUtils function in the Shift Details component has been
adjusted to return null or false values where necessary for properties
that depend on there being a shifts list with a size greater than 0.

The Cancel Shift Confirmation has been modified to not clear all shift
information in its afterConfirm function, since this also clears the
newly set shifts fetched after completing the DELETE operation, and
prematurely closes the modal (by unsetting the selected shift),
preventing the selection of the Close button and therefore the
triggering of the afterConfirm function.