Add incident report deletion from the Incident Reports panel
Added UI, server routes and functions, and apiUtil to support the deletion
of incident reports from the Incident Reports page of the Shift Details drawer.
On confirmation, the specified incident report will be deleted from the shift
document in the database as well as the Cloudinary file storage.

Also added a success alert component with state and an invokeCallback function
with error handling to the Confirmation component. The Confirmation can now
receive a successAlert as a prop, which will be displayed in the dialog on
successful confirmation if it remains open on confirm.

The error handling ensures that the callback function passed to the Confirmation
is interrupted, and an error alert message set, if the callback throws an error
during its execution.