Add incident report editing to Incident Report Details panel
Users can now enter 'edit mode' when viewing an individual
Incident Report, update the text content, and save it. The new
value will be saved in the database and displayed in the UI.

The existing PDF will be deleted from Cloudinary, and a new one
created and stored. Its PDF URL will also be updated in the database.