Log console messages rather than throwing errors for Cloudinary in shift controller
The storing and editing of notes in the application database should
not be dependent on the third-party Cloudinary service.

Having the note creation and editing server functions coupled with
Cloudinary API calls that could fail or throw errors potentially
prevents users from interacting with the notes in the app database.

These server functions have been modified to catch and log errors rather
than throwing them to the error handler and interrupting CRUD operations.