Add deletion of Incident Reports from the Incident Report Details view
Also modified the modal and drawer state and history update functions in
modalUtils to ensure that closing the modal does not have the side effect
of mutating the drawer state or history. prevDrawer has been renamed to
drawerHistory for clarity.

store.selectedShift has been added as a dependency to the useEffect in
Calendar which fetches and refreshes shift state, to ensure that modifying
the contents of a shift (notes, incidents, etc.) will also refresh the
list of shifts and the in-progress shift accessed by the 'View Shift In
Progress' button in the sidebar.