Fix bug causing clientID to be undefined in Add Shift Form under certain circumstances
Because the inital fetch of the client data returns their carers as a
list of strings (rather than objects), the clientID field in the Add Shift
Form could return undefined in cases where it was navigated to directly
after the Select Client view, since it looks for the id as an object
property rather than a string.

This has been remedied by calling getCarers and setting a list of carer
objects in state when loading the Calendar component, so that the carers
of the selectedClient is always a list of objects (or an empty list), rather
than a string when the Add Shift Form mounts.

It would be best to further standardise and document the format of resources
returned from API calls in future, to avoid such errors.