Add My Account router path and modal dialog component to BrowserRouter
Also refactor all Modals that are children of /calendar route to
directly return a Modal component rather than rendering Outlet in
a Modal. The Calendar now has Outlet at the top level of its return.

All Modal and Confirmation dialogs have been refactored to accept
a modalId (string) and can receive a hasEndpoint (boolean) prop.
The prop modalId is used to set the id in modalUtils for better
targeting of dialogs when opening, closing, or updating content.

Setting hasEndpoint to true indicates that the Modal/Confirmation
is rendered via a client router endpoint, and the 'close' handler
will now navigate back in the the browser's history to the route
where the Modal was opened.

Omit hasEndpoint (default) if the Modal has no dedicated route
and is defined locally in its calling component or is a standalone
component to  be imported.

modalUtils has also been refactored to account for these changes
and properly set the ModalContext.Provider state depending on user
interactions with modals, confirmations, and the drawer.