Add user account editing from the My Account dialog
The user can now update their firstName, lastName, email, and password
fields via the My Account dialog. Validation has been added to the
userController for correct password hashing.

The updateLoadState method of the Form component can now pass errors
as a second argument to the setParentIsLoading function. This is useful
if an errors array needs to be passed up the component tree in the event
that a form passes validation and submits, but the submitForm method
itself throws an error (e.g., if the reponse status code is not 200).

Components with a handleChildLoadState further up the tree can then
use this errors array, or pass it up in turn to their parent component
via the setParentIsLoading method, or some other method they have
received from a parent as a prop.

Also made some minor adjustments to Icon styles for a softer aesthetic.