Significantly refactor the Calendar component to improve data fetching flow
The handleSelectClient method of the Client component now fetches the client
data and sets it in the global context state so it is available when the
Calendar component loads.

A handleRefreshCalendar method has been added to the Calendar component to
set the client in its local state, get all client shifts, and set the
previous, in-progress, and upcoming shifts in global state before the isLoading
state of the Calendar is set to false.

This prevents the Calendar components from attempting to render before their
required data has been fetched. It is also a useful utility function for
refreshing the Calendar as needed, and can be passed to child components.

The findPreviousShift, findInProgressShift, and findNextUpcomingShift functions
have been extracted from the Calendar component and moved into calendarUtils.js
for better modularity.