Ensure Add Shift Form sets newly created shift as the selected shift
The previous version of the handleUpdateShifts function assumed that the
newly created shift would always be the last shift. This was incorrect,
since there could be other existing shifts after the date of the new shift
being created.

The shiftController has been updated to populate the carer data in the
newly created shift returned by the createShift route, and this is set
as the selected shift by handleUpdateShifts.

This ensures that selecting 'Manage shift' after creating a shift always
loads the newly-created shift, and not the last shift in the shifts array.