Add transitions when loading Shift Details sub-panels for smoother continuity
Also added a view property to the 'open' 'drawer' modalDispatch action that
explicitly sets the sub-panel that the drawer should open if provided.
If none is provided, it will open the Shift Overview by default.

Navigating to a sub-panel within the Shift Details Drawer will scroll
the drawer back to the top, however, this loses the scroll position
when navigating back to the Overview. A better solution may to be have
the sub-panels render as child routes and use the ScrollRestoration
component of react router to preserve the scroll restoration between
location changes.