Significantly refactor app router and add paths for all Shift Details Drawer views
The router location is now used to control Shift Details Drawer navigation rather than updating the activeDrawer property in modalUtils. Selecting the ‘back’ button navigates to the path with the last path segment sliced off, until the user reaches calendar/shift-details again.

Selecting a Shift from the Select Shift By Date modal no longer closes the modal when opening the drawer, since it may be convenient to return to the shifts of the selected date when closing the drawer again.

The ‘Home’ link in the main nav now displays ‘Clients’ when the user is logged in, to better indicate that it returns to the Client Selection view.

Some minor cosmetic changes have been made, such as the styling and placement of the ‘Coordinator’ label in the Care Team list in the sidebar of the Calendar view.

Most Modal components have been made direct children of the Calendar component, to make them accessible throughout the calendar view and shift details view one a client has been selected. The My Account modal appears in both the root path and in the Calendar, since it needs to be accessed from the Select Client view as well.

Some components which require data from a selected shift (such as the Edit Shift Form and Confirm Cancel Shift modal) have been made children of the Shift Details Drawer, but where possible modals should be rendered as children of the Calendar.

The hasEndpoint property has been removed from all components that return Modals currently in use, but has been left on the Modal and Confirmation components but may be removed in future. In general, avoid assigning endpoints to modals.

The Modal component now has an onClose prop which accepts a function to be executed when handling the closing of the modal. This is useful if closing a certain modal needs to force an update of some part of the app state that may not otherwise be updated (e.g., refreshing some data displayed in the UI after closing a confirmation that changed that data in the database).

For consistency, the Modal alert prop has been modified to take an object containing a severity (one of info, success, warning, or error) and a message, and the Modal component itself will take care of rendering the alert rather than every component which returns a modal having to set alerts as its children.

The alerts within a modal can also be positioned above or below the modal children by passing the strings ‘top’ or ‘bottom’ to the alertPosition prop of the modal.

Also for consistency, components which return a modal and contain buttons should pass those buttons to the Modal component via the actions prop where possible, to be displayed in the DialogActions component of the Modal's Dialog.