Replace all instances of setActiveDrawer with navigate
Since the active drawer and drawer history are now derived from the
router location, setActiveDrawer and drawerHistory have been removed
from modalUtils.

A navUtils property has been added to the global store, which can be
used to set globally accessible navigation utility functions to the
global context. Use the 'setNavUtil' dispatch action with a 'name' and
'function' prop to set a function in store.navUtils.

The handleViewShift function in the Shift Details Drawer has been
renamed handleNavigateToShift for clarity, and has also been extended
to accept a shift _id string or shift object. It has also been added
to store.navUtils as 'navigateToShift'.

Any component can now import the navigateToShift function from
store.navUtils and use it to navigate the drawer to a different shift
with the transition animation.