Add encryption keys to the frontend and backend servers
The encryption keys are derived from a password and salt stored
as environment variables. The symmetric key pair can be used to
encrypt and decrypt data that needs to be sent securely between
the frontend and backend (e.g., session data).

The Session model has been removed, because sessions will be
managed using the express-session and connect-mongodb-session
packages.

Utility functions have been added to generate and store the key
on both the frontend and backend to make it accessible to other
parts of the application.