Create a user session on login and post encrypted session data to the backend server
Utility functions have been added to the apiUtils of the frontend server
to encrypt and decrypt session data. When the user logs in successfully,
a new session is created by the sessionController and the initial state
of the global context store is encrypted and posted to the server for
storage in the database.

The encryption key and initialisation vector for the AES-GCM algorithm
are stored within an object in the global state context and can be accessed
via crypto.key and crypto.iv respectively after destructuring the 'crypto'
property from the global context value.

The express.raw middleware has been added to the server to parse octet-streams,
since the encrypted session data is sent to the server as an ArrayBuffer.