Set loading state of the App using the global context state reducer
Added dispatch actions 'setAppIsLoading', which updates the application loading state,
and 'updateStore', which overwrites only the passed global store properties rather than
overwriting the entire store. Also added the 'appIsLoading' property to the emptyStore
object, which defaults to ’true’.

updateStore is useful in instances where multiple properties of the global store need
to be updated in a single action.

Dispatching loading state changes to the store in this way (rather than passing down a
local state) prevents race conditions for child components that use the load state of
the parent app for rendering, in cases where the loading state is set by async functions.
This stops flashes of inconsistent component rendering when an async function that modifies
the loading state is still in the process of updating, but a child component that uses
said state has already begun rendering.

For example, the Home route conditionally renders a loading symbol, the login form, or the
Protected Route, depending on the 'appIsLoading' and 'isAuth' state of the application.
If the page is refreshed, the saved global state needs to be retrieved from the server,
and this would previously cause a flash of the login page before rendering the logged-in
view, because the Home component commenced rendering before the async function which fetched
the session state and correctly updated the app loading state was called.

Also changed the Select Client view path from the root path to '/clients' for clarity.