Prevent persisting isConfirmed state on Confirmation when it closes immediately on confirm
If a Confirmation does not have the 'stayOpenOnConfirm' (boolean) prop,
it does not need to set 'isConfirmed' to true at all, and simply closes.

A 'clearSelectedIncidentReport' action has been added to globalUtils to
easily unset the selected incident report (e.g., when returning to the
Incident Report overview list).

Also check whether 'editMode' prop is present before attempting to set it
to false, to avoid raising an error when submitting shift or handover notes
for the first time, since the non-edit-mode versions of these forms do not
receive any props.