Simplify functions updating the cleint's care team  when adding users to it
The CareTeamList and AddShiftForm components no longer store
separate copies of the client, carer, or coordinator objects in
local state, and always pull them from the global store for
consistency.

Since the 'refreshCalendar' action refreshes the shift and client data
automatically, fetching and setting the clients in the global state
when adding users to the client's care team is redunant, so 'setCarer'
dispatch actions have been removed from the methods and useEffects
of the AddShiftForm and CareTeamList components.

Also added a 'setDefaultCarer' method to the AddShiftForm component,
which takes a 'carerId' as an optional argument and sets the default
carer in the AddShiftForm to the passed carer, the first carer in the
client's list of carers, or an empty string (no selection), depending
on the application state when the method is called.