Add functionality to remove a client from the SelectClient view
FRONTEND
The removal of the client will be confirmed first in a dialog, and all associated shifts are also deleted.
If the removed client is the 'selectedClient' or 'prevSelectedClient', these properties are set to their
default (empty object) states in the global store to prevent attempting to load a deleted client when
navigating back to the Calendar.

A 'deleteClient' helper function has been added to apiUtils, and the 'getClientList' and 'cancelShift'
functions have been extracted from their components and placed in apiUtils for easier reuse.

A timeout delay of 200 ms has been added to the resetting of the Confirmation component's
'isConfirmed' and 'alert' states to await the modal closing transition and prevent flashes of
alternate state when closing a confirmation.

The 'handleSetNewClient' function of the AddClientForm now assigns an empty array to the 'carers'
and 'shifts' properties of the newly created client, to prevent them from being undefined when
navigating to the Calendar after creating a new client.

The AddClientForm handlers have been updated to correctly refresh the client list when a new
client is added.

The 'clients' property of the global store is now correctly initialised as an array.

The NavBar will now only display the 'Calendar' button if a 'selectedClient' or 'prevSelectedClient'
is set, to prevent the confusing UX of bouncing back and forth between SelectClient and Calendar
when attempting to access the Calendar without any selected/previously selected client.

Added icons to the 'Add Client' button and action buttons in AddClientForm, and replaced some
square icons with their rounded versions. Icons have been added to the confirm buttons of the
Remove Carer and Remove Client confirmations, for consistency with other destructive confirmation
dialog buttons.

The Form component now takes a 'buttonStartIcon' prop to add an icon to primary buttons in forms.
Icons have been added to other submission buttons throughout the app that were missing them.

BACKEND
Stop logging 'sessionID' in the backend server console when updating the user session.