Refactor Calendar to simplify 'handleRefreshCalendar' and only use client data from the store
The client is now always retrieved from the store, and not stored in
a separate local state. The 'handleRefreshCalendar' method has been
refactored to set all shifts in a single 'updateStore' dispatch action
for simplicity.

The useEffects that refresh the client and shift data have had their
dependencies adjusted to minimise the possiblity of infinte render
loops when changing the selected client.