Seed new demo accounts with an in-progress sample shift
Adjusted some copy, styling, and spacing, as well as adding start icons
to the 'log in' and 'start demo' buttons on the Home view.

The 'clients' property of the global context state object has been changed
back to initialising as an object rather than an array, since the client
list returned by the API is actually an object containing the property
'coordinator', whose value is an array of clients for which the user is the
coordinator, and the property 'carer' whose value is an array of clients
for which the user is the carer.