Add route, functions, and UI to enable user account deletion from My Account
A Confirmation will be displayed before deleting the user account.
All clients for whom the user is a coordinator, and their associated
shifts, will also be deleted, before finally deleting the user account.

The 'handleLogout' function defined in the NavBar component has been added
to the global store in the newly-created 'functions' property. This makes
the logout function centrally defined, easy to update, and easily accessible
by any component that needs to execute a logout, without having to implement
its own logout function.

Functions can be added to the global store using the 'addFunction' dispatch
action. Supply a 'name' property (String) and a 'data' property, which
should be a reference to the function. The function can then be accessed or
destructured from 'store.functions'.

Also made some minor adjustments to the sample shift notes and account deletion
dialog copy for increased clarity.