Enable uploading Shift Notes to the cloud if no shiftNotesPDF value is in the shiftNotes object
This also displays the response from the upload request in a Snackbar alert.
An 'uploadResource' function has been added to apiUtils for easy access, and
the ButtonUpload component has been modified to handle all required parameters
and return an alert object containing the result of the upload request.

The ref 'drawerRef' has been added to the 'PaperProps' prop of the Drawer component,
for easier access to the underlying MUI Paper component (which is the scrollable part
of the Drawer).

The Form component can now receive a 'formId' prop, which passes the given id to the
rendered HTML form component, and a Boolean 'scrollIntoView' prop, which, if true, will
scroll the form into view when it loads. This is useful for contexts in which a form is
fully/partially rendered below the visible viewport (e.g., when toggling into edit mode within
a shift notes panel on a mobile device).

The ShiftNotesForm, IncidentReportForm, and HandoverNotesForm components have
been given a 'formId' value and now scroll into view when loaded for a better UX.