Display a Welcome dialog when starting a new demo account
BACKEND
An 'isNewUser' Boolean field has been added to the user model and controller.
When loading the app for the first time with a new demo account, a Welcome dialog
is displayed if this flag is true. It is set to false when Welcome is dismissed.

FRONTEND
The Modal component has been adjusted to be open by default when logging into a
new demo account, so as to display the Welcome dialog.

The 'handleAfterDeleteUserAccount' method of the MyAccount component has been
made asynchronous, and the Confirmation component now awaits the resolution of a
Promise when executing an 'afterConfirm' callback function. This ensures the correct
control flow when deleting and logging out of an account, and of any 'afterConfirm'
functions passed to a Confirmation.