Refactor cancellation of sample shifts to check the 'isSample' property of shift objects
Permanently flagging this in the database, rather than setting it in the frontend,
prevents errors raised when attempting to access the calendar of a client with no shifts.

Also added a Boolean 'isSample' field to the shift, client, and user models to
easily identify sample entities seeded by a demo account.